refactor(src/users/repositories/users.repository.ts): refactor methods.

+++ src/users/repositories/users.repository.ts
@@ -3,6 +3,7 @@ import { PrismaService } from 'src/prisma/prisma.service'
 import { CreateUserDto } from '../dto/create-user.dto'
 import { UpdateUserDto } from '../dto/update-user.dto'
 import { UserEntity } from '../entities/user.entity'
+import { NotFoundError } from 'src/common/errors/types/notFoundError'
 
 @Injectable()
 export class UsersRepository {
@@ -15,7 +16,16 @@ export class UsersRepository {
   }
 
   async findAll(): Promise<UserEntity[]> {
-    return this.prisma.user.findMany()
+    return this.prisma.user.findMany({
+      include: {
+        posts: {
+          select: {
+            title: true,
+            created_at: true,
+          },
+        },
+      },
+    })
   }
 
   async findOne(id: number): Promise<UserEntity> {
@@ -23,19 +33,47 @@ export class UsersRepository {
       where: {
         id,
       },
+      include: {
+        posts: {
+          select: {
+            title: true,
+            created_at: true,
+          },
+        },
+      },
     })
   }
 
   async update(id: number, updateUserDto: UpdateUserDto): Promise<UserEntity> {
+    const user = await this.findOne(id)
+
+    if (!user) {
+      throw new NotFoundError('User not found!')
+    }
+
     return this.prisma.user.update({
       where: {
         id,
       },
       data: updateUserDto,
+      include: {
+        posts: {
+          select: {
+            title: true,
+            created_at: true,
+          },
+        },
+      },
     })
   }
 
   async remove(id: number): Promise<UserEntity> {
+    const user = await this.findOne(id)
+
+    if (!user) {
+      throw new NotFoundError('User not found!')
+    }
+
     return this.prisma.user.delete({
       where: {
         id,
